Do min update for lambda handler update
